`vault` package depends on `paas-vault-plugon-auth-cf`
Previously it was the other way around. This meant the package had to be
explicitly included in a manifest file, when in really only the `vault` package
should need to be included.

By doing it this way around, the `vault` package's dependency on
`paas-vault-plugin-auth-cf` will cause it's packaging to be run. The `vault`
package can then copy the binary during its packaging.