Capture absolute path of Vault plugin directory at runtime
Hashicorp Vault will not work with a plugin directory path that is a symlink.
It requires an absolute path.

Bosh packages under `/var/vcap/packages/*` are symlinks to
`/var/vcap/data/packages/*/*/`. We can't know the path when the configuration
file is generated, so we have to resort to resolving it at runtime.

To do that, we put a placeholer in the config file, use `readlink -f` to get
the absolute path of the symlink, and then use `sed` to replace it.

Monit starts the Vault process using the newly updated config file.